+++ config.py
@@ -1,17 +1,6 @@
-import numpy as np
+from helper import get_locations_centroids
 
-
-def get_locations_centroids(locations, key="coords"):
-
-    lats = []
-    longs = []
-    for loc, loc_data in locations.items():
-        coords = loc_data.get(key)
-        lats.append(coords[0])
-        longs.append(coords[1])
-    return np.mean(lats), np.mean(longs)
-
-
+# app configurations
 KEY_LOCATIONS = {
     "hang_xanh": {
         "name": "Hang Xanh Intersection",
@@ -21,8 +10,15 @@ KEY_LOCATIONS = {
         "name": "Tan Son Nhat International Airport",
         "coords": (10.8125471, 106.66566368265137),
     },
-    "bay_hien": {"name": "Bay Hien Intersection", "coords": (10.7928977, 106.653385)},
+    "bay_hien": {
+        "name": "Bay Hien Intersection",
+        "coords": (10.7928977, 106.653385),
+    },
 }
 
 LOCATION_CENTROID = get_locations_centroids(KEY_LOCATIONS)
 MAP_RADIUS = 4500
+
+# ui configurations
+INACTIVE_COLOR = "#A9A9A9"
+ACTIVE_COLOR = "#FF5733"
